Bounding box generation
Implemented generation of bounding boxes for everything in the cubes, spheres, and planes vectors within the first frame. everything after the first frame will need to have the bounding box made when it is created, and indexed properly. I'll add member functions to the shape structs to update the bounding boxes if the shape moves at a later time.